+++ src/Reader.php
@@ -377,7 +377,7 @@ class Reader extends AbstractCsv implements TabularDataReader, JsonSerializable
      */
     public function mapHeader(array $headers): TabularDataReader
     {
-        return Statement::create()->process($this, $headers);
+        return (new Statement())->process($this, $headers);
     }
 
     /**
@@ -388,7 +388,7 @@ class Reader extends AbstractCsv implements TabularDataReader, JsonSerializable
      */
     public function filter(Query\Predicate|Closure $predicate): TabularDataReader
     {
-        return Statement::create()->where($predicate)->process($this);
+        return (new Statement())->where($predicate)->process($this);
     }
 
     /**
@@ -400,7 +400,7 @@ class Reader extends AbstractCsv implements TabularDataReader, JsonSerializable
      */
     public function slice(int $offset, int $length = -1): TabularDataReader
     {
-        return Statement::create()->offset($offset)->limit($length)->process($this);
+        return (new Statement())->offset($offset)->limit($length)->process($this);
     }
 
     /**
@@ -411,7 +411,7 @@ class Reader extends AbstractCsv implements TabularDataReader, JsonSerializable
      */
     public function sorted(Query\Sort|Closure $orderBy): TabularDataReader
     {
-        return Statement::create()->orderBy($orderBy)->process($this);
+        return (new Statement())->orderBy($orderBy)->process($this);
     }
 
     /**
@@ -426,7 +426,7 @@ class Reader extends AbstractCsv implements TabularDataReader, JsonSerializable
      */
     public function matching(string $expression): iterable
     {
-        return FragmentFinder::create()->findAll($expression, $this);
+        return (new FragmentFinder())->findAll($expression, $this);
     }
 
     /**
@@ -440,7 +440,7 @@ class Reader extends AbstractCsv implements TabularDataReader, JsonSerializable
      */
     public function matchingFirst(string $expression): ?TabularDataReader
     {
-        return FragmentFinder::create()->findFirst($expression, $this);
+        return (new FragmentFinder())->findFirst($expression, $this);
     }
 
     /**
@@ -455,7 +455,7 @@ class Reader extends AbstractCsv implements TabularDataReader, JsonSerializable
      */
     public function matchingFirstOrFail(string $expression): TabularDataReader
     {
-        return FragmentFinder::create()->findFirstOrFail($expression, $this);
+        return (new FragmentFinder())->findFirstOrFail($expression, $this);
     }
 
     public function select(string|int ...$columns): TabularDataReader